backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
pillow==12.1.1
qrcode==8.2
sqlparse==0.5.5
